@@ -2,7 +2,7 @@ package gplay.marlonaguirre.ml.gplay.activities;
 
 import android.content.Intent;
 import android.os.Handler;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 
 import gplay.marlonaguirre.ml.gplay.R;
