@@ -2,11 +2,10 @@ package gplay.marlonaguirre.ml.gplay.adapters;
 
 import android.content.Context;
 import android.graphics.Bitmap;
-import android.graphics.drawable.Drawable;
 import android.net.Uri;
 import android.provider.MediaStore;
-import android.support.annotation.NonNull;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -16,7 +15,6 @@ import android.widget.TextView;
 import java.util.ArrayList;
 
 import gplay.marlonaguirre.ml.gplay.R;
-import gplay.marlonaguirre.ml.gplay.activities.MainActivity;
 import gplay.marlonaguirre.ml.gplay.pojos.Song;
 
 public class SongsAdapter extends RecyclerView.Adapter<SongsAdapter.ViewHolderSongs>
